requirements.txt
streamlit==1.26.0
pandas==1.5.3
numpy==1.23.5
seaborn==0.12.2
matplotlib==3.6.3
scipy==1.10.1
Pillow==9.4.0
openpyxl==3.1.2
